More tests, better forAllAsync error messages, actually run IOs returned from tests
